+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 ------
 
 * It is now possible to force the reinstallation of a plugin on `POST /plugins`
+  with query parameter `reinstall=True`.
 
 
 18.02